+++ CHANGELOG.md
@@ -51,3 +51,4 @@ and this project adheres to [Semantic Versioning](http://semver.org/).
 - EmailSender repeatable
 - Emails to BRIDGHEAD_ADMINS_WHO_HAVE_NOT_ACCEPTED_NOR_REJECTED_THE_PROJECT
 - FrontendSiteModule repeatable
+- Action names