+++ CHANGELOG.md
@@ -4,7 +4,7 @@ All notable changes to this project will be documented in this file.
 The format is based on [Keep a Changelog](http://keepachangelog.com/)
 and this project adheres to [Semantic Versioning](http://semver.org/).
 
-## [0.0.1 - 2023-12-30]
+## [0.0.1 - 2024-01-03]
 ### Added
 - First version of the project
 - Spring Application
@@ -55,3 +55,4 @@ and this project adheres to [Semantic Versioning](http://semver.org/).
 - Module names
 - Http Method and Params for Actions
 - Test endpoint
+- Download application form